pyproject.toml
[tool.poetry]
name = "gibs-imagestat"
version = "0.2.0a25"
description = "Application capable of computing statistics on GIBS images"
authors = ["podaac-tva <podaac-tva@jpl.nasa.gov>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/podaac/gibs-imagestat"
exclude = ['contrib', 'docs', 'tests']
packages = [
    { include = "imagestat" },
]

[tool.poetry.dependencies]
python = "^3.9"
urllib3 = "^1.26"
Pillow = "^9.4"
requests = "^2.28"
numpy = "^1.24"
pandas = "^1.4.1"
beautifulsoup4 = "^4.10.0"
starlette = "^0.25"
xarray = "^2023.1"
fastapi = "^0.92"
uvicorn = "^0.20"
lxml = "^4.9.1"

[tool.poetry.dev-dependencies]
pytest = "^7.2"
pylint = "^2.15.8"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
